backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# AI & ML
openai==1.3.5
chromadb==0.4.18
sentence-transformers==2.2.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==14.0.1
sqlalchemy==2.0.23
pymongo==4.6.0

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Handling
python-magic==0.4.27
Pillow==10.1.0

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
validators==0.22.0

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Monitoring
sentry-sdk==1.38.0

# Documentation
drf-spectacular==0.26.5

# Data Validation
cerberus==1.3.5
jsonschema==4.20.0

# Async Support
channels==4.0.0
channels-redis==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
